update stub (Laravel >5.8 generates bigint columns for ids by default)

+++ tests/database/migrations/0000_00_00_000000_create_tags_table.php
@@ -14,8 +14,8 @@ class CreateTagsTable extends Migration
     public function up()
     {
         Schema::create('tags', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('parent_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('parent_id')->nullable();
             $table->string('name');
             $table->unsignedInteger('position');
             $table->timestamps();
